organizing and creating tests for executions api
